requirements.txt
Django>=4.2,<5.0
django-mptt>=0.14.0
python-decouple>=3.8
Pillow>=10.0.0
